+++ composer.json
@@ -5,6 +5,7 @@
     "license": "MIT",
     "require": {
         "php": "^8.0",
+        "symfony/dependency-injection": "^5.4 || ^6.0",
         "symfony/http-kernel": "^5.4 || ^6.0",
         "symfony/property-access": "^5.4 || ^6.0",
         "twig/twig": "^3.4"
