+++ spec/Core/Twig/Extension/SortByExtensionSpec.php
@@ -11,7 +11,7 @@
 
 declare(strict_types=1);
 
-namespace spec\Sylius\Ux\Twig\Extension;
+namespace spec\Sylius\Ux\Core\Twig\Extension;
 
 use Doctrine\Common\Collections\ArrayCollection;
 use PhpSpec\ObjectBehavior;
