+++ tests/application/composer.json
@@ -1,0 +1,3 @@
+{
+    "name": "example/test-application"
+}
